Move users loading from UsersTable to CompanyUsersAdministration
